requirements.txt
numpy
pickle-mixin
streamlit
scikit-learn
pandas
requests
Pillow